bricks: add a component list
The event has a few complicated sensors which are not available usually
so create a new page to describe everything in detail about the places
from where you can get the sensor.

Note: This is incomplete as I have just created the backbone. This will
further be edited by Shivang.

Signed-off-by: Pranit Bauva <pranit.bauva@gmail.com>